ZCS-4221 Concrete implementation of getListeners for LocalImapMessageStore
The problem that this resolved is as follows:

The `closeFolder` method in `ImapSessionManager` asks the
`ImapMailboxStore` for listeners.  The implementation,
defined in `ImapMailboxStore`, worked correctly for remote
IMAP sessions, but not local ones.

This commit moved the implementation that was defined in
ImapMailboxStore to RemoteImapMailboxStore and added the
appropriate implementation to LocalImapMailboxStore.

Once this was done the closeFolder method could correctly detach
the appropriate sessions, which would caused them to be
unloaded from `sessions`.
